+++ lib/date.ex
@@ -3,6 +3,7 @@ defmodule DateTimeParser.Date do
 
   import NimbleParsec
   import DateTimeParser.Combinators.Date
+  import DateTimeParser.Formatters, only: [format_token: 2]
 
   defparsec(
     :parse,
@@ -17,4 +18,12 @@ defmodule DateTimeParser.Date do
     |> optional()
     |> concat(us_date())
   )
+
+  def from_tokens(tokens) do
+    Date.new(
+      format_token(tokens, :year) || 0,
+      format_token(tokens, :month) || 0,
+      format_token(tokens, :day) || 0
+    )
+  end
 end
